+++ champlain/champlain-view.h
@@ -131,7 +131,7 @@ gdouble champlain_view_longitude_to_x (ChamplainView *view,
 gdouble champlain_view_latitude_to_y (ChamplainView *view, 
     gdouble latitude);
 
-void champlain_view_get_viewport_origin (ChamplainView *view, gdouble *x, gdouble *y);
+void champlain_view_get_viewport_origin (ChamplainView *view, gint *x, gint *y);
 
 void champlain_view_bin_layout_add (ChamplainView *view,
   ClutterActor *child,
